Custom implementation of 234-tree for use with a student database
// Author: Dillon Johnson
// Date Started: December 2015
/* PROJECT
DESCRIPTION
* Creating a 2-3-4 tree from scratch and processing standard
commands via the cmd line
* Availible Commands:
*  - add(student
name,student_id)  // Adds a student structure to the tree. The student
name can contain any characters except for commas, parenthesis, and non
displayable characters. The Student ID is used as the value of sorting
whilst building the 234 tree itself.
*  - countLevels()      // Displays
an integer value indicating the number of levels in the tree, an empty
tree has 0 levels.
*  - displayInOrder()     // Displays the contents of
the tree in order of ascending student ID number.
*  -
getName(student_id)    // Searches the tree for the student ID provided,
if the name is found, the name is printed, else, print NOT FOUND.
*  -
getLevel(student_id)    // Finds the level on which the student with the
specified ID number resides and prints it to the user.
*  -
countValuesInNode(student_id)  // Finds the student with the specified
ID number and displays an integer representing the number of values that
reside within the node.
*  - quit()
*/